eslint complete

+++ client/components/companies/CompanyDelete.tsx
@@ -24,38 +24,39 @@ export const CompanyDelete = ({ company, setIsCompanyChanged, addNote, setShow, 
 
     const handleDelete = async (e: React.MouseEvent<HTMLButtonElement>) => {
         e.preventDefault()
-        const userConfirmed = window.confirm(`Willst du das Unternehmen '${company.data.name}' wirklich löschen?`)
-        if (userConfirmed) {
-            setIsLoading(true)
-            const client = await apiClient
-            client.deleteCompanyById(company.meta.location, null, { headers: { Authorization: `Bearer ${token}` } })
-                .then(
-                    result => {
-                        setIsLoading(false)
-                        setIsCompanyChanged(true)
-                        const note: Note = {
-                            variant: 'warning',
-                            message: `Unternehmen wurde gelöscht.`,
-                        }
-                        if (setShow) {
-                            setShow(false)
-                        }
-                        addNote(note)
-                        if (typeof result.headers.permissions === 'string') {
-                            updateUserPermissions(result.headers.permissions, permissions, setPermissions)
-                        } else {
-                            throw new Error ('No permissions header found')
-                        }
-                    },
-                    error => {
-                        setIsLoading(false)
-                        const note: Note = {
-                            variant: 'danger',
-                            message: `Löschen der Unternehmen hat nicht geklappt: ${error.message}`,
-                        }
-                        addNote(note)
+        if (token) {
+            const userConfirmed = window.confirm(`Willst du das Unternehmen '${company.data.name}' wirklich löschen?`)
+            if (userConfirmed) {
+                setIsLoading(true)
+                try {
+                    const client = await apiClient
+                    const result = await client.deleteCompanyById(company.meta.location, null, { headers: { Authorization: `Bearer ${token}` } })
+                    setIsLoading(false)
+                    setIsCompanyChanged(true)
+                    const note: Note = {
+                        variant: 'warning',
+                        message: `Unternehmen wurde gelöscht.`,
                     }
-                )
+                    if (setShow) {
+                        setShow(false)
+                    }
+                    addNote(note)
+                    if (typeof result.headers.permissions === 'string') {
+                        updateUserPermissions(result.headers.permissions, permissions, setPermissions)
+                    } else {
+                        throw Error('Permissions header should be type string.')
+                    }
+                } catch (error) {
+                    setIsLoading(false)
+                    const note: Note = {
+                        variant: 'danger',
+                        message: `Löschen der Unternehmen hat nicht geklappt: ${error instanceof Error ? error.message : String(error)}`,
+                    }
+                    addNote(note)
+                }
+            }
+        } else {
+            throw Error('Bitte authentifizieren')
         }
     }
     return <Button size={size}  variant='outline-danger' onClick={handleDelete}>Löschen</Button>
